# mycute-go/src/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "## API概要\\nMYCUTE REST APIを定義する。\\nURL最大長のリスクを避ける為、検索は query parameter ではなく body json を使用する。\\nGin に SEARCH/QUERY method の実装がないため、検索を POST にて行う。",
        "title": "MYCUTE",
        "contact": {}
    },
    "basePath": "/",
    "paths": {
        "/v1/chat_models/": {
            "post": {
                "description": "- USR によってのみ使用できる",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1 ChatModel"
                ],
                "summary": "ChatModel作成",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateChatModelParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CreateChatModelRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    }
                }
            }
        },
        "/v1/chat_models/search": {
            "post": {
                "description": "- USR によってのみ使用できる",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1 ChatModel"
                ],
                "summary": "ChatModelを検索",
                "parameters": [
                    {
                        "type": "string",
                        "example": "Bearer ??????????",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Search Params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SearchChatModelsParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/SearchChatModelsRes"
                        }
                    },
                    "400": {
                        "description": "Validation Error",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    }
                }
            }
        },
        "/v1/chat_models/{chat_model_id}": {
            "get": {
                "description": "- USR によってのみ使用できる",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1 ChatModel"
                ],
                "summary": "ChatModel詳細取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ChatModel ID",
                        "name": "chat_model_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetChatModelRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    }
                }
            },
            "delete": {
                "description": "- USR によってのみ使用できる",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1 ChatModel"
                ],
                "summary": "ChatModel削除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ChatModel ID",
                        "name": "chat_model_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DeleteChatModelRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    }
                }
            },
            "patch": {
                "description": "- USR によってのみ使用できる",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1 ChatModel"
                ],
                "summary": "ChatModel更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ChatModel ID",
                        "name": "chat_model_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateChatModelParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UpdateChatModelRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CreateChatModelParam": {
            "type": "object",
            "required": [
                "api_key",
                "model",
                "name",
                "provider"
            ],
            "properties": {
                "api_key": {
                    "type": "string",
                    "example": "sk-proj-..."
                },
                "base_url": {
                    "type": "string",
                    "example": "https://api.openai.com/v1"
                },
                "max_tokens": {
                    "type": "integer",
                    "example": 4096
                },
                "model": {
                    "type": "string",
                    "example": "gpt-4o"
                },
                "name": {
                    "type": "string",
                    "example": "My GPT-4"
                },
                "provider": {
                    "type": "string",
                    "example": "openai"
                },
                "temperature": {
                    "type": "number",
                    "example": 0.2
                }
            }
        },
        "CreateChatModelRes": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/CreateChatModelResData"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Err"
                    }
                }
            }
        },
        "CreateChatModelResData": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "DeleteChatModelRes": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/DeleteChatModelResData"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Err"
                    }
                }
            }
        },
        "DeleteChatModelResData": {
            "type": "object"
        },
...
        "EmptyObj": {
            "type": "object"
        },
        "Err": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 100000
                },
                "field": {
                    "type": "string",
                    "example": "field_name"
                },
                "message": {
                    "type": "string",
                    "example": "Some Error Message"
                }
            }
        },
        "ErrRes": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/EmptyObj"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Err"
                    }
                }
            }
        },
...
        "GetChatModelRes": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/GetChatModelResData"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Err"
                    }
                }
            }
        },
        "GetChatModelResData": {
            "type": "object",
            "properties": {
                "base_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "max_tokens": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "temperature": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "SearchChatModelsParam": {
            "type": "object",
            "properties": {
                "base_url": {
                    "type": "string",
                    "example": "https://api.openai.com/v1"
                },
                "model": {
                    "type": "string",
                    "example": "gpt-4o"
                },
                "name": {
                    "type": "string",
                    "example": "My GPT-4"
                },
                "provider": {
                    "type": "string",
                    "example": "openai"
                }
            }
        },
        "SearchChatModelsRes": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/SearchChatModelsResData"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Err"
                    }
                }
            }
        },
        "SearchChatModelsResData": {
            "type": "object",
            "properties": {
                "base_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "max_tokens": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "temperature": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "UpdateChatModelParam": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string",
                    "example": "sk-proj-..."
                },
                "base_url": {
                    "type": "string",
                    "example": "https://api.openai.com/v1"
                },
                "max_tokens": {
                    "type": "integer",
                    "example": 4096
                },
                "model": {
                    "type": "string",
                    "example": "gpt-4o"
                },
                "name": {
                    "type": "string",
                    "example": "My GPT-4"
                },
                "provider": {
                    "type": "string",
                    "example": "openai"
                },
                "temperature": {
                    "type": "number",
                    "example": 0.2
                }
            }
        },
        "UpdateChatModelRes": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/UpdateChatModelResData"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Err"
                    }
                }
            }
        },
        "UpdateChatModelResData": {
            "type": "object"
        },
...
    }
}